move Account, Money into model package
